+++ backend/src/main/java/com/c1331tjava/ServiceApp/controller/BidClientController.java
@@ -2,28 +2,18 @@ package com.c1331tjava.ServiceApp.controller;
 
 import com.c1331tjava.ServiceApp.config.SecurityConfig;
 import com.c1331tjava.ServiceApp.dto.BidDTO;
-import com.c1331tjava.ServiceApp.dto.BidDetailsDTO;
-import com.c1331tjava.ServiceApp.dto.ClientDTO;
 import com.c1331tjava.ServiceApp.exception.CrossUserException;
 import com.c1331tjava.ServiceApp.exception.CustomedHandler;
 import com.c1331tjava.ServiceApp.model.Bid;
 import com.c1331tjava.ServiceApp.model.Request;
-import com.c1331tjava.ServiceApp.model.UserEntity;
 import com.c1331tjava.ServiceApp.service.BidService;
 import com.c1331tjava.ServiceApp.service.RequestService;
 import com.c1331tjava.ServiceApp.service.UserEntityService;
-import jakarta.validation.Valid;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
 
 /**
  * Class to handle Client side REST request of Bid.class
@@ -47,8 +37,13 @@ public class BidClientController {
      * @param id: Id of requested bid details
      * @return BidDTO with details of a Bid filtered by Provider and Request
      */
+
+    @Operation(summary = "Get Bid details by id",
+            description = "<p>Only authorized to Client users.</p>" +
+                    "<p>Extract query username from JWT.</p>" +
+                    "<p>A user can only query about bids of owned request. </p>")
     @GetMapping("/details/{id}")
-    public BidDTO findByProviderAndRequest(@Valid @PathVariable Long id){
+    public BidDTO findByProviderAndRequest(@PathVariable Long id){
 
         String currentUser = securityConfig.getUserNameFromToken();
         Bid currentBid;
@@ -57,33 +52,19 @@ public class BidClientController {
         // Query Bid by Id and the Request it belongs
         try {
             currentBid = this.bidService.findById(id).get();
+        } catch (Exception e) {
+            throw new CustomedHandler("Error retrieving data from Bid database");
+        }
+        try {
             currentRequest = this.requestService.findById(currentBid.getRequest().getId()).get();
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new CustomedHandler("Error retrieving data from Request database");
         }
+
         // If the user owner of the request differ to currentuser, throw cross user exception
         if (!currentRequest.getClient().getEmail().equals(currentUser)){
             throw new CrossUserException();
         }
         return modelMapper.map(currentBid, BidDTO.class);
-    }
-
-    /**
-     *
-     * Class to handle @Valid exception and personalize error message
-     *
-     * @param ex type: MethodArgumentNotValidException, exception thrown by @Valid
-     * @return Personalized error message
-     */
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public Map<String, String> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getAllErrors().forEach((error) -> {
-            String fieldName = "Bad Request";
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        return errors;
     }
 }